+++ Tests/StencilTests/InheritenceSpec.swift
@@ -11,17 +11,17 @@ func testInheritence() {
 
     $0.it("can inherit from another template") {
       let template = try environment.loadTemplate(name: "child.html")
-      try expect(try template.render()) == "Header\nChild"
+      try expect(try template.render()) == "Super_Header Child_Header\nChild_Body"
     }
 
     $0.it("can inherit from another template inheriting from another template") {
       let template = try environment.loadTemplate(name: "child-child.html")
-      try expect(try template.render()) == "Child Child Header\nChild"
+      try expect(try template.render()) == "Super_Header Child_Header Child_Child_Header\nChild_Body"
     }
 
     $0.it("can inherit from a template that calls a super block") {
       let template = try environment.loadTemplate(name: "child-super.html")
-      try expect(try template.render()) == "Header\nChild Body"
+      try expect(try template.render()) == "Header\nChild_Body"
     }
   }
 }